Initial commit for testing input pins
Using the cloudiuno library, this code is testing
the HIGH and LOW signals on PIN 0 with a 6 seconds
delay.

You need to use a volt-meter to check if the voltage
is changing after 6 seconds.

Bear in mind that here the HIGH means low voltage and
LOW means high voltage which is just an observation.